fix: cleaning up debugging and code
I've cleaned up a bunch of the debugging so the logs are clearer now. I've narrowed the timeout problem down to the generated certificates. Something about them is preventing the connection from being fully established. Likely rejected in some way but as for why is inscrutable.

More digging to follow. But I know that If I load the cert files from memory then it works. So I know it's the generated certs.

* Related #8

[ci skip]