fix: addressed crashing due to socket send failures
This commit addresses the send failures by taking any send specific errors and passing them back to the connection to be handled.
Any errors such as bad arguments results in the connection throwing the problem proactivity.
Any network failures are generally ignored and the Connections are left to time out. This allows for the network to drop for a short amount of time without failure of the connection and streams.

[ci skip]