fix: fixed edge cases leading to failures in ci
There were a few things here. For starters some of the new tests had some dumb config. A mistake on my part.

The other part is that the `send` throws different errors depending on platform and context. I could only reasonably track it down based on testing in the CI... There are 3 main errors we're sending back to the connection to be dealt with.