wip: small fixes
socketCleanupFactory maintains a set of sockets it tracks. So it will hold sockets in memory. It's not needed since its only purpose is to clean up sockets during tests if they fail the test.

We want to track each connection and stream on the server side and make sure they finish before moving to the next step of the loop.